backend/requirements.txt
# Python dependencies for Flask backend
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
supabase==2.4.3
pandas==2.2.2
edu-convokit==0.1.0
gunicorn==22.0.0
werkzeug==3.0.3
openai>=1.50.0
